+++ model_navigator/api/utilities.py
@@ -23,10 +23,10 @@ from typing import Any, Callable, Optional
 
 from model_navigator.api.config import Framework, OptimizationProfile, SizedDataLoader
 from model_navigator.commands.find_max_batch_size.find_max_batch_size import MaxBatchSizeFinder
+from model_navigator.core.dataloader import to_numpy
 from model_navigator.core.tensor import FRAMEWORK_TO_TENSOR_TYPE, PyTreeMetadata, TensorMetadata
 from model_navigator.exceptions import ModelNavigatorProfilingError
 from model_navigator.runners.registry import get_runner
-from model_navigator.utils.dataloader import to_numpy
 
 logger_name = "model_navigator.api.utilities"
 LOGGER = logging.getLogger(logger_name)