+++ model_navigator/commands/convert/converters/ts2onnx.py
@@ -19,8 +19,8 @@ from typing import Any, Dict, List, Optional
 import fire
 import torch  # pytype: disable=import-error
 
+from model_navigator.core.dataloader import load_samples
 from model_navigator.core.tensor import TensorMetadata
-from model_navigator.utils.dataloader import load_samples
 
 
 def convert(