+++ model_navigator/commands/export/exporters/torch2torchscript.py
@@ -20,9 +20,9 @@ import fire
 import torch  # pytype: disable=import-error
 
 from model_navigator.api.config import JitType
+from model_navigator.core.dataloader import load_samples
 from model_navigator.core.tensor import TensorMetadata
 from model_navigator.exceptions import ModelNavigatorUserInputError
-from model_navigator.utils.dataloader import load_samples
 
 
 def get_model() -> torch.nn.Module: