+++ model_navigator/commands/find_max_batch_size/find_max_batch_size_script.py
@@ -20,10 +20,10 @@ import fire
 
 from model_navigator.api.config import OptimizationProfile
 from model_navigator.commands.find_max_batch_size.find_max_batch_size import MaxBatchSizeFinder
+from model_navigator.core.dataloader import load_samples
 from model_navigator.core.logger import LOGGER
 from model_navigator.core.tensor import TensorMetadata
 from model_navigator.runners.registry import get_runner
-from model_navigator.utils.dataloader import load_samples
 
 
 def find_max_batch_size(