+++ model_navigator/commands/performance/profile_script.py
@@ -20,9 +20,9 @@ import fire
 
 from model_navigator.api.config import OptimizationProfile
 from model_navigator.commands.performance.profiler import Profiler
+from model_navigator.core.dataloader import load_samples
 from model_navigator.core.tensor import TensorMetadata
 from model_navigator.runners.registry import get_runner
-from model_navigator.utils.dataloader import load_samples
 
 
 def get_model() -> object: