+++ tests/functional/L0_Optimize_and_Save_jax_linear/test.sh
@@ -19,7 +19,7 @@ THIS_SCRIPT_PATH="$(realpath --relative-to="$(pwd)" "$0")"
 TEST_MODULE="$(dirname "${THIS_SCRIPT_PATH}"|sed 's/\//./g').test"
 
 #JAX_VERSION="jax[cuda11_pip]"
-JAX_VERSION="jax[cuda12_pip]"
+JAX_VERSION="jax[cuda12_pip]<0.4.16"
 
 pip install --upgrade ${JAX_VERSION} -f https://storage.googleapis.com/jax-releases/jax_cuda_releases.html
 pip install flax